trying to implement reliable branching

+++ src/encoding/UniformTimeDiscretization.h
@@ -51,7 +51,7 @@ ProblemEncoding<Field> to_uniform_time_milp(const STN<Field>& problem,
             std::tuple{&unit, task, t},
             builder.new_variable(
                 std::format("Q({}, {}, {})", unit.get_id(), task->get_id(), t),
-                VariableType::REAL, 0, info.batch_max_size));
+                VariableType::INTEGER, 0, info.batch_max_size));
         starts.emplace(
             std::tuple{&unit, task, t},
             builder.new_variable(
